Slightly reorganize build modules
