+++ crates/relayer-components-extra/src/components/extra/closures/mod.rs
@@ -1,4 +1,3 @@
 pub mod batch;
-pub mod build;
 pub mod chain;
 pub mod relay;
