+++ crates/relayer-components/src/components/default/closures/mod.rs
@@ -1,3 +1,2 @@
-pub mod build;
 pub mod relay;
 pub mod transaction;
